Renamed userModel.ts to UserModel.ts and todoModel.ts to TodoModel.ts
